+++ spec/concepts/stellar_base/withdrawal_requests/contracts/create_spec.rb
@@ -1,20 +1,15 @@
-require 'spec_helper'
+require "spec_helper"
 
 module StellarBase
   module WithdrawalRequests
     module Contracts
       RSpec.describe Create do
-
         describe "validations" do
           %i[
             asset_type
             asset_code
             dest
-            issuer
-            account_id
             fee_network
-            fee_fixed
-            fee_percent
           ].each do |attr|
             it "requires #{attr}" do
               contract = described_class.new(WithdrawalRequest.new)
@@ -22,8 +17,16 @@ module StellarBase
               expect(contract.errors[attr]).to include("can't be blank")
             end
           end
-        end
 
+          context "valid asset_code" do
+            it "requires inclusion in withdrawable_assets" do
+              contract = described_class.new(WithdrawalRequest.new)
+              contract.validate(asset_code: "BTCH")
+              expect(contract.errors[:asset_code])
+                .to include "invalid asset_code. Valid asset_codes: BTCT"
+            end
+          end
+        end
       end
     end
   end