+++ spec/requests/withdraw_spec.rb
@@ -1,17 +1,17 @@
 require "spec_helper"
 
-describe "GET /withdraw", type: :request, vcr: {record: :once} do
+describe "GET /withdraw", type: :request, vcr: { record: :once } do
   let(:uri) do
     StellarBase::Engine.routes.url_helpers.withdraw_path
   end
 
   before do
     @withdraw_request = @bridge_callback = nil
-    StellarBase.configuration.on_withdraw = ->(withdrawal_request, bridge_callback) do
+    StellarBase.configuration.on_withdraw = -> (withdrawal_request, bridge_callback) do
       @withdrawal_request = withdrawal_request
       @bridge_callback = bridge_callback
     end
-    StellarBase.configuration.on_bridge_callback = ->(bridge_callback) do
+    StellarBase.configuration.on_bridge_callback = -> (bridge_callback) do
       StellarBase::WithdrawalRequests::Process.(bridge_callback)
     end
   end
@@ -43,16 +43,16 @@ describe "GET /withdraw", type: :request, vcr: {record: :once} do
 
     it "executes the given instructions" do
       # NOTE: see WithdrawalRequestRepresenter spec for detailed specs
-      expect(json_response[:account_id]).
-        to eq StellarBase.configuration.distribution_account
+      expect(json_response[:account_id])
+        .to eq StellarBase.configuration.distribution_account
       expect(json_response[:memo]).to be_present
 
       # Execute bridge callback
       StellarBase::BridgeCallbacks::Process.(bridge_callback)
 
       expect(@withdrawal_request).to be_present
-      expect(@withdrawal_request.account_id).
-        to eq StellarBase.configuration.distribution_account
+      expect(@withdrawal_request.account_id)
+        .to eq StellarBase.configuration.distribution_account
       expect(@withdrawal_request.asset_type).to eq "crypto"
       expect(@withdrawal_request.asset_code).to eq "BTCT"
       expect(@withdrawal_request.issuer).to eq CONFIG[:issuer_address]
@@ -64,6 +64,27 @@ describe "GET /withdraw", type: :request, vcr: {record: :once} do
       expect(@withdrawal_request.memo).to be_present
 
       expect(@bridge_callback).to eq bridge_callback
+    end
+  end
+
+  context "payment for an invalid asset" do
+    it "returns an error" do
+      get(uri, {
+        params: {
+          type: "crypto",
+          asset_code: "BCHT",
+          dest: "my-bcht-address",
+          fee_network: 0.001,
+          format: :json,
+        },
+      })
+
+      expect(response).not_to be_success
+      json = JSON.parse(response.body).with_indifferent_access
+
+      expect(json["error"]).to be_present
+      expect(json["error"]["asset_code"])
+        .to include "invalid asset_code. Valid asset_codes: BTCT"
     end
   end